notification coordinate float -> int
I know next to nothing about coding, but this seemed to have done the trick for me

After updating to the newest Anki release 25.07.5, The Exam Notifier produced an error message everytime the pop-up was supposed to appear. AI pointed this out to me so please don't expect any actual understanding on my part. 

Qt6 seems to not accept the float values anymore, so casting them to integers could fix this issue (at least it did for me up until now)